Refactored stand-alone functions to separate scanimage_utils module
